backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
psycopg==3.2.12
psycopg-binary==3.2.12
pydantic==2.11.10
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.1
websockets==15.0.1
